tests/playwright/requirements.txt
playwright>=1.40.0
pytest-playwright>=0.4.3
pytest-asyncio>=0.23.0
requests>=2.31.0
paramiko>=3.4.0
docker>=7.0.0
